Update nrf9160 macOS install
Nordic changed cask name from `nordic-nrf5x-command-line-tools` to `nordic-nrf-command-line-tools`, so the install failed. Also moved the cask install to a separate line. We should discuss splitting the brew install with many formulas into separate commands, as reading the output and finding failed installs in the massive output is difficult and error-prone.